+++ swift/extractor/incompatible-extractor.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+exec -a "$0" "$(dirname $0)/incompatible-os"
