+++ swift/ql/integration-tests/autobuilder/unsupported-os/diagnostics.expected
@@ -1,9 +1,9 @@
 {
-  "markdownMessage": "Currently, `autobuild` for Swift analysis is only supported on macOS.\n\n[Change the Actions runner][1] to run on macOS.\n\nYou may be able to run analysis on Linux by setting up a [manual build command][2].\n\n[1]: https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idruns-on\n[2]: https://docs.github.com/en/enterprise-server/code-security/code-scanning/automatically-scanning-your-code-for-vulnerabilities-and-errors/configuring-the-codeql-workflow-for-compiled-languages#adding-build-steps-for-a-compiled-language",
+  "markdownMessage": "Currently, Swift analysis is only supported on macOS.\n\n[Change the Actions runner][1] to run on macOS.\n\n[1]: https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idruns-on",
   "severity": "error",
   "source": {
     "extractorName": "swift",
-    "id": "swift/autobuilder/incompatible-os",
+    "id": "swift/extractor/incompatible-os",
     "name": "Incompatible operating system (expected macOS)"
   },
   "visibility": {
