+++ swift/tools/autobuild.sh
@@ -5,5 +5,5 @@ if [[ "$OSTYPE" == "darwin"* ]]; then
   export CODEQL_SWIFT_POD_EXEC=`which pod`
   exec "${CODEQL_EXTRACTOR_SWIFT_ROOT}/tools/${CODEQL_PLATFORM}/swift-autobuilder"
 else
-  exec "${CODEQL_EXTRACTOR_SWIFT_ROOT}/tools/${CODEQL_PLATFORM}/autobuilder-incompatible-os"
+  exec "${CODEQL_EXTRACTOR_SWIFT_ROOT}/tools/${CODEQL_PLATFORM}/incompatible-os"
 fi
