+++ swift/tools/incompatible-os/BUILD.bazel
@@ -1,8 +1,8 @@
 load("//swift:rules.bzl", "swift_cc_binary")
 
 swift_cc_binary(
-    name = "autobuilder-incompatible-os",
-    srcs = ["AutobuilderIncompatibleOs.cpp"],
+    name = "incompatible-os",
+    srcs = ["IncompatibleOs.cpp"],
     # No restrictions (Windows allowed)
     target_compatible_with = [],
     visibility = ["//swift:__subpackages__"],
